requirements.txt
# Core ML/DL libraries
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.3.0

# Data processing and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# PDF parsing
PyMuPDF>=1.20.0
pdfminer.six>=20221105

# Utilities
tqdm>=4.64.0
tensorboard>=2.10.0
Pillow>=9.0.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional: GPU support
# cuda-python>=11.7.0  # Uncomment if using CUDA
